+++ packages/rs-sdk-ffi/src/data_contract/queries/info.rs
@@ -1,0 +1,1 @@
+
