+++ packages/rs-sdk-ffi/src/data_contract/queries/mod.rs
@@ -1,0 +1,2 @@
+mod fetch;
+mod info;
