+++ packages/rs-sdk-ffi/src/document/create.rs
@@ -130,15 +130,8 @@ mod tests {
     use crate::test_utils::test_utils;
     use crate::test_utils::test_utils::*;
     use crate::DashSDKErrorCode;
-    use dash_sdk::dpp::data_contract::accessors::v1::DataContractV1Getters;
-    use dash_sdk::dpp::data_contract::conversion::json::DataContractJsonConversionMethodsV0;
-    use dash_sdk::dpp::data_contract::document_type::DocumentTypeRef;
-    use dash_sdk::dpp::data_contract::v1::DataContractV1;
-    use dash_sdk::dpp::document::document_factory::DocumentFactory;
     use dash_sdk::dpp::identity::{Identity, IdentityV0};
-    use dash_sdk::dpp::platform_value::BinaryData;
     use dash_sdk::dpp::prelude::Identifier;
-    use dash_sdk::dpp::version::PlatformVersion;
     use std::collections::BTreeMap;
     use std::ffi::{CStr, CString};
     use std::ptr;
