+++ packages/rs-sdk-ffi/src/document/queries/info.rs
@@ -8,40 +8,6 @@ use crate::sdk::SDKWrapper;
 use crate::types::{DashSDKDocumentInfo, DocumentHandle, SDKHandle};
 use crate::{DashSDKError, DashSDKErrorCode, FFIError};
 
-/// Destroy a document
-#[no_mangle]
-pub unsafe extern "C" fn dash_sdk_document_destroy(
-    sdk_handle: *mut SDKHandle,
-    document_handle: *mut DocumentHandle,
-) -> *mut DashSDKError {
-    if sdk_handle.is_null() || document_handle.is_null() {
-        return Box::into_raw(Box::new(DashSDKError::new(
-            DashSDKErrorCode::InvalidParameter,
-            "Invalid parameters".to_string(),
-        )));
-    }
-
-    let wrapper = &mut *(sdk_handle as *mut SDKWrapper);
-    let _document = &*(document_handle as *const Document);
-
-    let result: Result<(), FFIError> = wrapper.runtime.block_on(async {
-        // Use DocumentDeleteTransitionBuilder to delete the document
-        // We need to get the data contract and document type information
-        // This is a simplified implementation - in practice you might need more context
-
-        // For now, return not implemented as we need more context about the data contract
-        Err(FFIError::InternalError(
-            "Document deletion requires data contract context - use specific delete function"
-                .to_string(),
-        ))
-    });
-
-    match result {
-        Ok(_) => std::ptr::null_mut(),
-        Err(e) => Box::into_raw(Box::new(e.into())),
-    }
-}
-
 /// Get document information
 #[no_mangle]
 pub unsafe extern "C" fn dash_sdk_document_get_info(
@@ -98,12 +64,4 @@ pub unsafe extern "C" fn dash_sdk_document_get_info(
     };
 
     Box::into_raw(Box::new(info))
-}
-
-/// Destroy a document handle
-#[no_mangle]
-pub unsafe extern "C" fn dash_sdk_document_handle_destroy(handle: *mut DocumentHandle) {
-    if !handle.is_null() {
-        let _ = Box::from_raw(handle as *mut Document);
-    }
 }
